refactor: modularize `handleResponse`
